+++ src/Cake.Board.Abstractions/IWorkItem.cs
@@ -36,9 +36,7 @@ namespace Cake.Board.Abstractions
         /// <summary>
         /// Add work item to release note.
         /// </summary>
-        /// <typeparam name="T">Concrete impelmentation of <see cref="IWorkItem"/>.</typeparam>
-        /// <param name="releaseNote">Reference to release note.</param>
-        void ToReleaseNotes<T>(ref IReleaseNotes<T> releaseNote)
-            where T : IWorkItem;
+        /// <returns>A string rappresenting the content of work item.</returns>
+        string ToReleaseNotes();
     }
 }