+++ src/Cake.Board.AzureBoards/Models/WorkItem.cs
@@ -1,10 +1,9 @@
 // Copyright (c) Nicola Biancolini, 2019. All rights reserved.
 // Licensed under the MIT license. See the LICENSE file in the project root for full license information.
 
-using System.Linq;
+using System;
+
 using Cake.Board.Abstractions;
-using Cake.Board.Extensions;
-using Newtonsoft.Json;
 
 namespace Cake.Board.AzureBoards.Models
 {
@@ -34,15 +33,6 @@ namespace Cake.Board.AzureBoards.Models
         public string Url { get; set; }
 
         /// <inheritdoc/>
-        public void ToReleaseNotes<T>(ref IReleaseNotes<T> releaseNote)
-            where T : IWorkItem
-        {
-            releaseNote.NotNull(nameof(releaseNote));
-
-            if (this.Type == "Bug" && this.Type == "Issue")
-                releaseNote.BugFixes.ToList().Add((T)this);
-            else
-                releaseNote.Enhancements.ToList().Add(this);
-        }
+        public string ToReleaseNotes() => $"- [**#{this.Id.TrimStart('#')}**]({this.Url}): {this.Title}  {Environment.NewLine}{this.Description}";
     }
 }