+++ test/Cake.Board.AzureBoards.Tests/Specs/GenerateReleaseNotesSpec.cs
@@ -1,7 +1,9 @@
 // Copyright (c) Nicola Biancolini, 2019. All rights reserved.
 // Licensed under the MIT license. See the LICENSE file in the project root for full license information.
 
+using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Threading.Tasks;
 
 using Cake.Board.AzureBoards.Commands;
@@ -13,7 +15,8 @@ namespace Cake.Board.AzureBoards.Tests.Specs
 {
     public class GenerateReleaseNotesSpec
     {
-        private IEnumerable<WorkItem> _wits;
+        private readonly IEnumerable<WorkItem> _wits;
+        private readonly string _releaseNotes;
 
         public GenerateReleaseNotesSpec()
         {
@@ -21,57 +24,103 @@ namespace Cake.Board.AzureBoards.Tests.Specs
             {
                 new WorkItem
                 {
-                    Id = "10",
+                    Id = "120",
                     Type = "Bug",
-                    Url = "https://lorem.ipsum/10",
-                    Title = "in tellus integer feugiat scelerisque varius morbi enim nunc faucibus",
+                    Url = "https://lorem.ipsum/120",
+                    Title = "URL parts incorrectly identifies the user and repository parts.",
                     Description = string.Empty,
                     State = "Closed"
                 },
                 new WorkItem
                 {
-                    Id = "8",
+                    Id = "119",
                     Type = "Issue",
-                    Url = "https://lorem.ipsum/8",
-                    Title = "risus feugiat in ante metus dictum at tempor commodo ullamcorper",
+                    Url = "https://lorem.ipsum/119",
+                    Title = "Correct issue with casing of label name.",
                     Description = string.Empty,
                     State = "Closed"
                 },
                 new WorkItem
                 {
-                    Id = "100",
+                    Id = "124",
                     Type = "Task",
-                    Url = "https://lorem.ipsum/100",
-                    Title = "nam at lectus urna duis convallis convallis tellus id interdum",
+                    Url = "https://lorem.ipsum/124",
+                    Title = "Add support for adding default labels to issues on a repository.",
+                    Description = "Create a class to insert labels on issue from configuration file.",
+                    State = "Closed"
+                },
+                new WorkItem
+                {
+                    Id = "117",
+                    Type = "Feature",
+                    Url = "https://lorem.ipsum/117",
+                    Title = "Add Token option for CLI as alternative to user/password.",
                     Description = string.Empty,
                     State = "Closed"
                 },
                 new WorkItem
                 {
-                    Id = "1",
+                    Id = "117",
                     Type = "Feature",
-                    Url = "https://lorem.ipsum/1",
-                    Title = "eu nisl nunc mi ipsum faucibus vitae aliquet nec ullamcorper",
+                    Url = "https://lorem.ipsum/117",
+                    Title = "Add Token option for CLI as alternative to user/password.",
+                    Description = string.Empty,
+                    State = "Closed"
+                },
+                new WorkItem
+                {
+                    Id = "116",
+                    Type = "Epic",
+                    Url = "https://lorem.ipsum/116",
+                    Title = "Add GitReleaseManager as Global Tool",
+                    Description = string.Empty,
+                    State = "Closed"
+                },
+                new WorkItem
+                {
+                    Id = "121",
+                    Type = "Epic",
+                    Url = "https://lorem.ipsum/121",
+                    Title = "Include link to closed issues in milestone link.",
+                    Description = string.Empty,
+                    State = "Closed"
+                },
+                new WorkItem
+                {
+                    Id = "115",
+                    Type = "Feature",
+                    Url = "https://lorem.ipsum/115",
+                    Title = "Switch to new csproj format.",
+                    Description = string.Empty,
+                    State = "Closed"
+                },
+                new WorkItem
+                {
+                    Id = "110",
+                    Type = "Task",
+                    Url = "https://lorem.ipsum/110",
+                    Title = "Update project NuGet packages",
                     Description = string.Empty,
                     State = "Closed"
                 }
             };
+            this._releaseNotes = File.ReadAllText($"{Environment.CurrentDirectory}/release-notes.md");
         }
 
         [Fact(DisplayName = @"GIVEN a DevOps engineer with a list of work items
 WHEN he wants to create a release notes
 THEN it must be able to obtain the content sought")]
         [Trait(TraitNames.TEST_CATEGORY, TraitValues.ACCEPTANCE_TEST)]
-        public async Task ScenarioFromBoardExtension_SearchWorkItemByQueryId()
+        public void ScenarioFromBoardExtension_SearchWorkItemByQueryId()
         {
             // Arrange
             var fakeCakeContext = new FakeCakeContext(logBehaviour: () => new FakeCakeLog());
 
             // Act
-            await fakeCakeContext.GenerateReleaseNotesAsync(default, this._wits);
+            string releaseNotes = fakeCakeContext.GenerateReleaseNotes(this._wits);
 
             // Assert
-            Assert.True(false);
+            Assert.Equal(this._releaseNotes, releaseNotes);
         }
     }
 }