+++ test/Cake.Board.AzureBoards.Tests/release-notes.md
@@ -1,3 +1,12 @@
-# {title}
-## {wit_type}
-- [**#{wit.id}**]({wit.url}): {wit.title}
+# Bugs and issues.
+- [**#119**](https://lorem.ipsum/119): Correct issue with casing of label name.
+- [**#120**](https://lorem.ipsum/120): URL parts incorrectly identifies the user and repository parts.
+
+# Enhancements
+- [**#110**](https://lorem.ipsum/110): Update project NuGet packages
+- [**#115**](https://lorem.ipsum/115): Switch to new csproj format.
+- [**#116**](https://lorem.ipsum/116): Add GitReleaseManager as Global Tool
+- [**#117**](https://lorem.ipsum/117): Add Token option for CLI as alternative to user/password.
+- [**#121**](https://lorem.ipsum/121): Include link to closed issues in milestone link.
+- [**#124**](https://lorem.ipsum/124): Add support for adding default labels to issues on a repository.  
+Create a class to insert labels on issue from configuration file.